pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sage"
version = "0.1.0"
description = "Sage - Simplified AI agent definition and deployment via configuration"
authors = [
    {name = "Sage Choi", email = "sachoi@apollo.com"}
]
requires-python = ">=3.10,<3.13"
readme = "README.md"
dependencies = [
    "litellm>=1.40",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "aiofiles>=24.0",
    "aiosqlite>=0.20",
    "click>=8.0",
    "numpy>=1.26",
    "mcp>=1.0",
    "apollo-logging>=0.3.0",
    "python-dotenv>=1.0",
    "textual>=0.50",
    "httpx>=0.27",
    "markdownify>=0.14",
    "tomli>=2.0; python_version < '3.11'",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1.0",
    "coverage>=7.3.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "tomli>=2.0",
]

[project.scripts]
sage = "sage.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["sage"]

[[tool.uv.index]]
name = "apollo"
url = "https://pkgs.dev.azure.com/ApolloAzureDevOps/_packaging/ApolloAzureDevOps/pypi/simple/"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=sage --cov-report=term-missing  --cov-report=xml"

[tool.coverage.run]
source = ["sage"]
omit = [
    "tests/*",
    "*/tests/*",
    "*/__pycache__/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
